pyproject.toml
[project]
name = "langchain-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["aiofiles==25.1.0",
    "aiogram==3.23.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "attrs==25.4.0",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "distro==1.9.0",
    "frozenlist==1.8.0",
    "groq==0.37.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.0",
    "langchain-core==1.2.5",
    "langchain-groq==1.1.1",
    "langchain-openai==1.1.6",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.1",
    "langsmith==0.5.1",
    "magic-filter==1.0.12",
    "multidict==6.7.0",
    "openai==2.14.0",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==25.0",
    "propcache==0.4.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "uuid-utils==0.12.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",]
